refactor: remove type.ts file's types because it's supposed to be in the domain
