+++ src/domains/canvasInterface.ts
@@ -1,8 +1,3 @@
-import { Position } from '@/types'
-import ColorThief from 'colorthief'
-import { HTMLCanvas } from './dom/HTMLCanvas'
-const colorThief = new ColorThief()
-
 export interface CanvasInterface {
   isDrawnMask: boolean
   scale: number
