+++ src/domains/datasets.ts
@@ -1,17 +1,5 @@
 import { Dataset } from './dataset'
-import { DatasetInterface } from './datasetInterface'
-
-export type Position = {
-  xPx: number
-  yPx: number
-}
-export type Plot = {
-  id: number
-  xPx: number
-  yPx: number
-}
-
-export type Plots = Plot[]
+import { DatasetInterface, Plots } from './datasetInterface'
 
 export class Datasets {
   datasets: DatasetInterface[]
