+++ src/types/type.ts
@@ -1,14 +1,3 @@
-export type Position = {
-  xPx: number
-  yPx: number
-}
-
-export type Plot = { id: number } & Position
-
-export type Plots = Plot[]
-
-export type PlotValue = { xV: string; yV: string } & Plot
-
 export type DiameterRange = {
   min: number
   max: number
